Extract story-view from stories-view
